% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-functions.R
\name{grapes-or-or-grapes}
\alias{grapes-or-or-grapes}
\alias{\%||\%}
\title{NULL-coalescing operator}
\usage{
lhs \%||\% rhs
}
\arguments{
\item{lhs}{Left-hand side.}

\item{rhs}{Right-hand side.}
}
\value{
One of the operands.
}
\description{
Returns \code{lhs} if it is non-\code{NULL}, otherwise \code{rhs}. Lifted from
common R idioms to keep argument-defaulting code readable.
}
\keyword{internal}
